Made MEM-AP access work consistently
PSoC Programmer runs the "Acquire PSoC" sequence twice, so we do that
here as well. The "sticky overrun detection" code is borrowed from the
CMSIS-DAP driver, and is crucial to getting MEM-AP access to work. This
is all really hacky, but it seems to work. Except reset--that triggers a
SIGSEGV.